This change is a partial fix for the problem of long cluster failovers.  This change addresses part of that problem: long newbie-broker catchup times.
At this time, it looks as though there are two distinct mechanisms causing these long catchup times.  They roughly divide into a population of times less than 20 seconds, and a population of much longer times (up to hundreds of seconds).

This fix addresses only the shorter times.  In a test of 25 failovers before and after the change, it reduced those times by an average of nearly 50%.  A T-test on the times indicates that the likelihood of the observed change happening randomly is less than 2%.


git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk@954895 13f79535-47bb-0310-9956-ffa450edef68