Aggregate Timer warnings.
The Timer code logs a warning if a timer callback is started late or
overruns the start time for the next callback. In cases where there
are a lot of these warnings, the time taken to do the logging itself
severly worsens the situation.

This commit aggregates timer warnings and give a statistical report
every 5 seconds at most.

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk@1036169 13f79535-47bb-0310-9956-ffa450edef68