QPID-3775: Automate the manual persistent store tests.
Add new testcase BDBBackupTest to test the operation of the BDB store backup mechanism (backup.sh).
Changed test case PersistentStoreTest to implement the manually scripted BDB tests.
This required changes to QBTC and BrokerHolder to allow an external Brokers to be forcibly kill'd with -9.
Remove script and class for manual tests.

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk@1236931 13f79535-47bb-0310-9956-ffa450edef68